+++ hurocon/meta.py
@@ -1,11 +1,2 @@
-import sys
-
-
-if sys.version_info[:2] >= (3, 8):
-    from importlib import metadata
-else:
-    import importlib_metadata as metadata
-
-
-version = metadata.version(__package__)
+version = '0.3.0.dev0'
 author = 'maximilionus'