+++ CHANGELOG.md
@@ -12,13 +12,16 @@ This project uses [semantic versioning](https://semver.org/)
 ### Changed
 - Project structure enhanced to be more modular
 
+### Fixed
+- Misspelling issues in output
+
 
 ## 0.2.0 : 2022.05.03
 ### Added
 - Now any valid `url` address can be used to connect to router, not just `ip`
 
 ### Changed
-- Authentification password now will be stored more safely
+- Authentication password now will be stored more safely
 - Configuration file updated to version `2` - automatic merge supported
 - License changed to [MIT](./LICENSE)